move fake resources out
