+++ packages/dashboard/src/components/dashboard/tests/commands-panel.test.tsx
@@ -3,8 +3,8 @@ import * as RomiCore from '@osrf/romi-js-core-interfaces';
 import React from 'react';
 import { DeliveryRequestForm, LoopRequestForm } from 'react-components';
 import ResourceManager from '../../../managers/resource-manager';
+import fakeResources from '../../../managers/__mocks__/resources';
 import fakeFleets from '../../../mock/data/fleets';
-import fakeResources from '../../../mock/data/resources';
 import { ResourcesContext } from '../../app-contexts';
 import CommandsPanel from '../commands-panel';
