+++ packages/dashboard/src/tests/resource-manager-dispensers.test.ts
@@ -2,7 +2,7 @@ import {
   DispenserResource,
   DispenserResourceManager,
 } from '../managers/resource-manager-dispensers';
-import fakeResources from '../mock/data/resources';
+import fakeResources from '../managers/__mocks__/resources';
 
 describe('The correct operation of the getIconPath method of the DispenserResourceManager class', () => {
   let manager: DispenserResourceManager;
@@ -20,6 +20,7 @@ describe('The correct operation of the getIconPath method of the DispenserResour
 
   test('Returns falsy if the icon path it`s empty', () => {
     resourceData.testDispenser = {
+      guid: 'testDispenser',
       icons: {
         testDispenser: '',
       },
